fix: remove v1 suffix, shift some codes
