+++ src/lib/pages/account-details/types.ts
@@ -1,4 +1,11 @@
-import { type Option, type TokenWithValue, type Validator } from "lib/types";
+import { z } from "zod";
+
+import {
+  type Option,
+  type TokenWithValue,
+  type Validator,
+  zAddr,
+} from "lib/types";
 
 import type {
   Delegation,
@@ -32,3 +39,22 @@ export interface UserDelegationsData {
   totalCommission: Option<Record<string, TokenWithValue>>;
   isLoadingTotalCommission: boolean;
 }
+
+export enum TabIndex {
+  Overview = "overview",
+  Assets = "assets",
+  Delegations = "delegations",
+  Txs = "txs",
+  Codes = "codes",
+  Contracts = "contracts",
+  Admins = "admins",
+  Resources = "resources",
+  Modules = "modules",
+  Proposals = "proposals",
+}
+
+export const zAccDetailQueryParams = z.object({
+  accountAddress: zAddr,
+  // Remark: Didn't use enum here because we want to allow for redirect to overview
+  tab: z.string().default(TabIndex.Overview),
+});
